eth2/requirements.txt
web3
python-dotenv
eth-keys